fix:修复编译时incompatible types错误
类型不匹配

+++ src/main/java/com/webank/ddcms/service/impl/CompanyServiceImpl.java
@@ -75,7 +75,7 @@ public class CompanyServiceImpl implements CompanyService {
         companyInfoMapper.listCompanyWithStatus(
             Integer.parseInt(request.getAccountStatus()),
             request.getKeyWord(),
-            offset,
+            (long) offset,
             request.getPageSize());
     pageListData.setItemList(boList);
